chore(deploy): pass commit sha to production deployment
This commit adds the `COMMIT_SHA` environment variable to the context
of the production deployment job.

This makes the exact Git commit hash of the release available to the
on-server script. It is a preparatory step to allow the script to
perform a `git checkout` in the future, ensuring on-server files are
in sync with the deployed version.